backend/requirements.txt
arrow==1.0.3
asgiref==3.3.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
binaryornot==0.4.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cookiecutter==1.7.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
Django==3.1.7
django-cors-headers==3.7.0
django-db-views==0.1.1
djangorestframework==3.12.2
drf-yasg==1.20.0
greenlet==1.0.0
gunicorn==20.0.4
h11==0.12.0
haversine==2.3.0
httpcore==0.12.3
httpx==0.16.1
idna==2.10
importlib-metadata==3.10.0
importlib-resources==4.1.1
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jinja2-time==0.2.0
MarkupSafe==1.1.1
mysql==0.0.2
mysqlclient==2.0.3
numpy==1.20.2
openapi-codec==1.3.2
packaging==20.9
pandas==1.2.3
pkginfo==1.6.1
poyo==0.5.0
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2021.1
requests==2.25.1
rfc3986==1.4.0
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sniffio==1.2.0
soupsieve==2.2.1
SQLAlchemy==1.4.4
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
zipp==3.4.1
zstandard==0.15.1
